+++ python/ray/rllib/offline/__init__.py
@@ -8,6 +8,7 @@ from ray.rllib.offline.json_writer import JsonWriter
 from ray.rllib.offline.output_writer import OutputWriter, NoopOutput
 from ray.rllib.offline.input_reader import InputReader
 from ray.rllib.offline.mixed_input import MixedInput
+from ray.rllib.offline.shuffled_input import ShuffledInput
 
 __all__ = [
     "IOContext",
@@ -17,4 +18,5 @@ __all__ = [
     "OutputWriter",
     "InputReader",
     "MixedInput",
+    "ShuffledInput",
 ]
