+++ python/ray/rllib/offline/mixed_input.py
@@ -6,10 +6,10 @@ import numpy as np
 
 from ray.rllib.offline.input_reader import InputReader
 from ray.rllib.offline.json_reader import JsonReader
-from ray.rllib.utils.annotations import override, PublicAPI
+from ray.rllib.utils.annotations import override, DeveloperAPI
 
 
-@PublicAPI
+@DeveloperAPI
 class MixedInput(InputReader):
     """Mixes input from a number of other input sources.
 
@@ -21,7 +21,7 @@ class MixedInput(InputReader):
         }, ioctx)
     """
 
-    @PublicAPI
+    @DeveloperAPI
     def __init__(self, dist, ioctx):
         """Initialize a MixedInput.
